Use UNKNOWN state for failure to validate assertions
Assertions (file, registry) are currently hard-coded values
within the plugin source code. Because any errors in their
content is the fault of the plugin, we use an UNKNOWN state
instead of a WARNING or CRITICAL state (because the problem
is internal instead of external against the monitored
"reboot needed" values).

refs GH-93