Drop Summary section from output
The goal is to aggressively trim duplicated details from
output in order to make room for unique "reboot required"
evidence details and the Summary section duplicates
details already available in the one-line summary or
lead-in text.

Leaving the writeReportHeader helper function intact for
now with a nolint entry noting the applicable GH issue
for later consideration. Perhaps we'll add a flag
later to restore the output if we expand its purpose
(e.g., if we list the number of registry matches
alongside the number of file/directory matches).

refs GH-119